+++ str/ft_free_strs_array.c
@@ -6,16 +6,17 @@
 /*   By: chlimous <chlimous@student.42.fr>	    +#+  +:+	   +#+	      */
 /*						  +#+#+#+#+#+	+#+	      */
 /*   Created: 2025/04/20 19:50:40 by chlimous	       #+#    #+#	      */
-/*   Updated: 2025/04/20 20:05:02 by chlimous         ###   ########.fr       */
+/*   Updated: 2025/04/20 20:20:46 by chlimous         ###   ########.fr       */
 /*									      */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stddef.h>
 #include <stdlib.h>
 
 void	ft_free_strs_array(char ***array)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (array[i])
